% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{dynfeature}
\alias{dynfeature}
\title{Feature Importance for Dynamic Processes}
\description{
Calculating feature importance scores from trajectories using the random forests algorithm.
}
