% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_gene_importances.R
\name{trajectory_feature_importances}
\alias{trajectory_feature_importances}
\title{Calculating feature importances across trajectories}
\usage{
trajectory_feature_importances(traj, method = "ranger",
  method_params = list())
}
\arguments{
\item{traj}{A trajectory object containing expression values and a trajectory.}

\item{method}{The method to do regressions, can be `ranger` or any regression model from caret}

\item{method_params}{Parameters given to the method}
}
\description{
Uses the feature importance measures of \code{\link[ranger]{ranger}} or  \code{caret}.
}
