% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_feature_importances.R
\name{calculate_feature_importances}
\alias{calculate_feature_importances}
\title{Calculate feature importance scores}
\usage{
calculate_feature_importances(X, Y, fi_method = fi_ranger_rf_lite(),
  verbose = FALSE)
}
\arguments{
\item{X}{A data frame containing the features as columns.}

\item{Y}{A data frame of predictor variables, with \code{nrow(Y) == nrow(X)}.}

\item{fi_method}{A feature importance method. Default: \code{fi_ranger_rf_lite()}. Check \code{?fi_methods} for a full list of available feature importance methods.}

\item{verbose}{Whether to print out extra information.}
}
\description{
Calculate feature importance scores
}
